Remove unused legacy code and introduce session management
