+++ src/Argon.Api/Grains.Interfaces/ISessionManager.cs
@@ -1,0 +1,22 @@
+namespace Argon.Api.Grains.Interfaces;
+
+using MemoryPack;
+
+[Alias("Argon.Api.Grains.Interfaces.ISessionManager")]
+public interface ISessionManager : IGrainWithGuidKey
+{
+    [Alias("Authorize")]
+    Task<JwtToken> Authorize(UserCredentialsInput input);
+
+    [Alias("GetUser")]
+    Task GetUser();
+
+    [Alias("Logout")]
+    Task Logout(); // TODO: revoke jwt by adding it into a blacklist
+}
+
+[Serializable]
+[GenerateSerializer]
+[MemoryPackable]
+[Alias(nameof(JwtToken))]
+public partial record struct JwtToken(string token);
