--- src/Argon.Api/Grains.Persistence.States/ChannelType.cs
@@ -1,11 +1,0 @@
-namespace Argon.Api.Grains.Persistence.States;
-
-[GenerateSerializer]
-[Serializable]
-[Alias(nameof(ChannelType))]
-public enum ChannelType
-{
-    Text,
-    Voice,
-    Announcement
-}
