--- src/Argon.Api/Grains.Persistence.States/ServerChannelsStore.cs
@@ -1,12 +1,0 @@
-namespace Argon.Api.Grains.Persistence.States;
-
-using MemoryPack;
-
-[GenerateSerializer]
-[Serializable]
-[MemoryPackable]
-[Alias(nameof(ServerChannelsStore))]
-public sealed partial record ServerChannelsStore
-{
-    [Id(0)] public List<Guid> Channels { get; private set; } = [];
-}
