--- src/Argon.Api/Grains.Persistence.States/ServerToUserRelations.cs
@@ -1,12 +1,0 @@
-namespace Argon.Api.Grains.Persistence.States;
-
-using MemoryPack;
-
-[GenerateSerializer]
-[Serializable]
-[MemoryPackable]
-[Alias(nameof(ServerToUserRelations))]
-public sealed partial record ServerToUserRelations
-{
-    [Id(0)] public List<UserToServerRelation> Users { get; private set; } = [];
-}
