--- src/Argon.Api/Grains.Persistence.States/UserToServerRelations.cs
@@ -1,12 +1,0 @@
-namespace Argon.Api.Grains.Persistence.States;
-
-using MemoryPack;
-
-[GenerateSerializer]
-[Serializable]
-[MemoryPackable]
-[Alias(nameof(UserToServerRelations))]
-public sealed partial record UserToServerRelations
-{
-    [Id(0)] public List<UserToServerRelation> Servers { get; private set; } = [];
-}
